Add username field to /auth/token requests
Requests sent directly to the /auth/token endpoint end up having the username and password logged and eventually stored in the logs. The added filters grab the name while removing the decoded password. Further work to be done to remove the encoded request from all parts to scrub the password from the system while also allowing a username to be attached to a token.